# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/v1/mailTemplates": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mailTemplate"
                ],
                "summary": "Creates mail templates",
                "parameters": [
                    {
                        "description": "Mail template information",
                        "name": "mailTemplate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.MailTemplateForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/v1/needies": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "needy"
                ],
                "summary": "Lists all needies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "needy"
                ],
                "summary": "Creates needy",
                "parameters": [
                    {
                        "description": "Needy information",
                        "name": "needy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.NeedyForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/v1/needies/informations": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "needy"
                ],
                "summary": "Lists all needies informations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.MailTemplateForm": {
            "type": "object",
            "properties": {
                "mailTemplateType": {
                    "type": "integer"
                },
                "template": {
                    "type": "string"
                }
            }
        },
        "controllers.NeedyForm": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "object",
                    "$ref": "#/definitions/models.Address"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "needyCategories": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "phoneNumber": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "summary": {
                    "type": "string"
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        },
        "models.Address": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "firstLine": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "postalCode": {
                    "type": "string"
                },
                "secondLine": {
                    "type": "string"
                }
            }
        }
    }
}